Updated Audhie's ADTs

+++ adt/imp/point.c
@@ -5,35 +5,35 @@
 #define PI 3.14159265
 /* *** DEFINISI PROTOTIPE PRIMITIF *** */
 /* *** Konstruktor membentuk POINT *** */
-POINT MakePOINT (float X, float Y){
+POINT MakePOINT (int X, int Y){
 /* Membentuk sebuah POINT dari komponen-komponennya */
-POINT temp;
+	POINT temp;
 	Absis(temp)=X;
 	Ordinat(temp)=Y;
 	return temp;
 }
-/* *** KELOMPOK Interaksi dengan I/O device, BACA/TULIS  *** */                                                 
-void BacaPOINT (POINT * P){ 
-/* Membaca nilai absis dan ordinat dari keyboard dan membentuk 
+/* *** KELOMPOK Interaksi dengan I/O device, BACA/TULIS  *** */
+void BacaPOINT (POINT * P){
+/* Membaca nilai absis dan ordinat dari keyboard dan membentuk
    POINT P berdasarkan dari nilai absis dan ordinat tersebut */
 /* Komponen X dan Y dibaca dalam 1 baris, dipisahkan 1 buah spasi */
-/* Contoh: 1 2 
+/* Contoh: 1 2
    akan membentuk POINT <1,2> */
 /* I.S. Sembarang */
 /* F.S. P terdefinisi */
-	float a,b;
-	scanf("%f %f",&a,&b);
+	int a,b;
+	scanf("%d %d",&a,&b);
 	*P=MakePOINT(a,b);
 }
 void TulisPOINT (POINT P){
-/* Nilai P ditulis ke layar dengan format "(X,Y)" 
-   tanpa spasi, enter, atau karakter lain di depan, belakang, 
-   atau di antaranya 
+/* Nilai P ditulis ke layar dengan format "(X,Y)"
+   tanpa spasi, enter, atau karakter lain di depan, belakang,
+   atau di antaranya
    Output X dan Y harus dituliskan dalam bilangan riil dengan 2 angka di belakang koma.
 */
 /* I.S. P terdefinisi */
-/* F.S. P tertulis di layar dengan format "(X,Y)" */                
-printf("(%.2f,%.2f)",Absis(P),Ordinat(P));
+/* F.S. P tertulis di layar dengan format "(X,Y)" */
+	printf("(%d,%d)",Absis(P),Ordinat(P));
 }
 /* *** Kelompok operasi relasional terhadap POINT *** */
 boolean EQ (POINT P1, POINT P2){
@@ -67,19 +67,19 @@ int Kuadran (POINT P){
 		return 2;
 	else if ((Absis(P)<0)&&(Ordinat(P)<0))
 		return 3;
-	else 
+	else
 		return 4;
 }
-/* *** KELOMPOK OPERASI LAIN TERHADAP TYPE *** */                           
+/* *** KELOMPOK OPERASI LAIN TERHADAP TYPE *** */
 POINT NextX (POINT P){
 /* Mengirim salinan P dengan absis ditambah satu */
 	return (MakePOINT((Absis(P)+1),Ordinat(P)));
-}              
+}
 POINT NextY (POINT P){
 /* Mengirim salinan P dengan ordinat ditambah satu */
 	return(MakePOINT(Absis(P),(Ordinat(P)+1)));
 }
-POINT PlusDelta (POINT P, float deltaX, float deltaY){
+POINT PlusDelta (POINT P, int deltaX, int deltaY){
 /* Mengirim salinan P yang absisnya adalah Absis(P) + deltaX dan ordinatnya adalah Ordinat(P) + deltaY */
 	return((MakePOINT((Absis(P)+deltaX),(Ordinat(P)+deltaY))));
 }
@@ -102,7 +102,7 @@ float Panjang (POINT P1, POINT P2){
 /* Menghitung jarak terdekat antara P1 dan P2*/
 	return(sqrt(sqr(Absis(P1)-Absis(P2))+sqr(Ordinat(P1)-Ordinat(P2))));
 }
-void Geser (POINT *P, float deltaX, float deltaY){
+void Geser (POINT *P, int deltaX, int deltaY){
 /* I.S. P terdefinisi */
 /* F.S. P digeser, absisnya sebesar deltaX dan ordinatnya sebesar deltaY */
 	*P=PlusDelta(*P,deltaX,deltaY);
@@ -128,7 +128,7 @@ void Mirror (POINT *P, boolean SbX){
 /* Jika SbX false maka dicerminkan terhadap sumbu Y */
 	*P=MirrorOf(*P,SbX);
 }
-void Putar (POINT *P, float Sudut){
+void Putar (POINT *P, int Sudut){
 /* I.S. P terdefinisi */
 /* F.S. P digeser sebesar Sudut derajat dengan sumbu titik (0,0) */
 	float temp;
